refactor: merge board confirmation dialog
